+++ src/main/kotlin/net/trilleo/mc/plugins/tritown/commands/moderation/ReloadCommand.kt
@@ -4,7 +4,6 @@ import net.trilleo.mc.plugins.tritown.Main
 import net.trilleo.mc.plugins.tritown.registration.PluginCommand
 import net.trilleo.mc.plugins.tritown.utils.sendPrefixed
 import org.bukkit.command.CommandSender
-import org.bukkit.entity.Player
 import org.bukkit.plugin.java.JavaPlugin
 
 /**
@@ -21,19 +20,11 @@ class ReloadCommand(private val plugin: JavaPlugin) : PluginCommand(
     override fun execute(sender: CommandSender, args: Array<out String>): Boolean {
         val main = plugin as? Main
         if (main == null) {
-            if (sender is Player) {
-                sender.sendPrefixed("<red>Error: Plugin instance type mismatch. Unable to reload configuration.")
-            } else {
-                sender.sendMessage("Error: Plugin instance type mismatch. Unable to reload configuration.")
-            }
+            sender.sendPrefixed("<red>Error: Plugin instance type mismatch. Unable to reload configuration.")
             return true
         }
         main.reload()
-        if (sender is Player) {
-            sender.sendPrefixed("Configuration reloaded!")
-        } else {
-            sender.sendMessage("Configuration reloaded!")
-        }
+        sender.sendPrefixed("Configuration reloaded!")
         return true
     }
 }